+++ src/main/java/gs/mclo/api/APIException.java
@@ -5,8 +5,15 @@ import org.jetbrains.annotations.ApiStatus;
 import java.io.IOException;
 
 public final class APIException extends IOException {
+    private final int httpStatusCode;
+
     @ApiStatus.Internal
-    public APIException(String message) {
+    public APIException(String message, int httpStatusCode) {
         super(message);
+        this.httpStatusCode = httpStatusCode;
+    }
+
+    public int getHttpStatusCode() {
+        return httpStatusCode;
     }
 }