+++ src/test/java/gs/mclo/api/ApiTest.java
@@ -6,7 +6,7 @@ public class ApiTest {
     protected MclogsClient client;
 
     @BeforeEach
-    protected void setUp() {
+    protected void setUp() throws InterruptedException {
         client = new MclogsClient("aternos/mclogs-java-tests");
 
         var apiUrl = System.getenv("MCLOGS_API_URL");
@@ -15,5 +15,7 @@ public class ApiTest {
         if ((apiUrl != null && !apiUrl.isBlank()) || (logViewUrl != null && !logViewUrl.isBlank())) {
             client.setInstance(new Instance(apiUrl, logViewUrl));
         }
+
+        Thread.sleep(50); // Avoid rate limiting
     }
 }